// migrations/snapshots/12344505d11c7c5fb15a0040feb9e63338460f7106b7a66030686a4c87a07ef7/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Attachments: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly taskId: CodecTypes['pg/uuid@1']['output'];
      readonly uploadedById: CodecTypes['pg/uuid@1']['output'];
      readonly fileName: Varchar<255>;
      readonly storageKey: Varchar<255>;
      readonly mimeType: Varchar<100>;
      readonly fileSize: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly AuditLogs: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly taskId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly changedColumn: Varchar<100>;
      readonly oldValue: CodecTypes['pg/text@1']['output'] | null;
      readonly newValue: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly ProjectInvitations: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly projectId: CodecTypes['pg/uuid@1']['output'];
      readonly email: Varchar<255>;
      readonly invitedById: CodecTypes['pg/uuid@1']['output'];
      readonly tokenHash: Varchar<64>;
      readonly status: 'PENDING' | 'ACCEPTED' | 'EXPIRED' | 'CANCELED';
      readonly expiresAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly acceptedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly ProjectMembers: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly projectId: CodecTypes['pg/uuid@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Projects: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<150>;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly clientName: Varchar<150> | null;
      readonly status: 'ACTIVE' | 'COMPLETED' | 'ARCHIVED' | 'PLANNING';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly TaskDependencies: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly dependentTaskId: CodecTypes['pg/uuid@1']['output'];
      readonly dependencyTaskId: CodecTypes['pg/uuid@1']['output'];
      readonly createdBy: CodecTypes['pg/uuid@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Tasks: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly projectId: CodecTypes['pg/uuid@1']['output'];
      readonly assignedToId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly title: Varchar<200>;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly status: 'TODO' | 'BLOCKED' | 'IN_PROGRESS' | 'DONE';
      readonly priority: 'LOW' | 'MEDIUM' | 'HIGH' | 'URGENT';
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly clientVisible: CodecTypes['pg/bool@1']['output'];
      readonly version: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly Users: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: Varchar<100>;
      readonly email: Varchar<255>;
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly role: 'PM' | 'INTERNAL' | 'CLIENT';
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Attachments: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly taskId: CodecTypes['pg/uuid@1']['input'];
      readonly uploadedById: CodecTypes['pg/uuid@1']['input'];
      readonly fileName: CodecTypes['sql/varchar@1']['input'];
      readonly storageKey: CodecTypes['sql/varchar@1']['input'];
      readonly mimeType: CodecTypes['sql/varchar@1']['input'];
      readonly fileSize: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly AuditLogs: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly taskId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly changedColumn: CodecTypes['sql/varchar@1']['input'];
      readonly oldValue: CodecTypes['pg/text@1']['input'] | null;
      readonly newValue: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly ProjectInvitations: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly projectId: CodecTypes['pg/uuid@1']['input'];
      readonly email: CodecTypes['sql/varchar@1']['input'];
      readonly invitedById: CodecTypes['pg/uuid@1']['input'];
      readonly tokenHash: CodecTypes['sql/varchar@1']['input'];
      readonly status: 'PENDING' | 'ACCEPTED' | 'EXPIRED' | 'CANCELED';
      readonly expiresAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly acceptedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly ProjectMembers: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly projectId: CodecTypes['pg/uuid@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Projects: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly clientName: CodecTypes['sql/varchar@1']['input'] | null;
      readonly status: 'ACTIVE' | 'COMPLETED' | 'ARCHIVED' | 'PLANNING';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly TaskDependencies: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly dependentTaskId: CodecTypes['pg/uuid@1']['input'];
      readonly dependencyTaskId: CodecTypes['pg/uuid@1']['input'];
      readonly createdBy: CodecTypes['pg/uuid@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Tasks: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly projectId: CodecTypes['pg/uuid@1']['input'];
      readonly assignedToId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly title: CodecTypes['sql/varchar@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly status: 'TODO' | 'BLOCKED' | 'IN_PROGRESS' | 'DONE';
      readonly priority: 'LOW' | 'MEDIUM' | 'HIGH' | 'URGENT';
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly clientVisible: CodecTypes['pg/bool@1']['input'];
      readonly version: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly deletedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly Users: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['sql/varchar@1']['input'];
      readonly email: CodecTypes['sql/varchar@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly role: 'PM' | 'INTERNAL' | 'CLIENT';
      readonly department: 'PRODUCT' | 'UI_UX' | 'FRONTEND' | 'BACKEND' | 'CLIENT';
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;